nixos-tataru: update composer extension updater and stuff

+++ nixos-tataru/composer-extensions/update-composer-extension.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env nix-shell
-#!nix-shell -i bash -p jq phpPackages.composer nix-update
+#!nix-shell -i bash -p jq phpPackages.composer nix-update gnused
 
 ext=$1
 if [ -z "$ext" ]; then
@@ -15,13 +15,26 @@ echo $details | jq
 
 composerLock=$(echo $details | jq -r .composerLock)
 gitRepoUrl=$(echo $details | jq -r .gitRepoUrl)
+currentRev=$(echo $details | jq -r .currentRev)
+fileWithSrc=$(echo $details | jq -r .fileWithSrc)
+preferredBranch=$(echo $details | jq -r .preferredBranch)
 
-tmpdir=$(mktemp -d --suffix=composer-lock-updater)
+tmpdir=$(mktemp -d --suffix=-composer-lock-updater)
 cd $tmpdir
-git clone --recursive --depth 1 "$gitRepoUrl" $ext
+git clone --recursive --branch "$preferredBranch" --depth 1 "$gitRepoUrl" $ext
 cd $ext
+newRev=$(git rev-parse HEAD)
+if [ $currentRev = $newRev ]; then
+	set +x
+	echo "$ext is already updated ($currentRev)"
+	cd $curdir
+	rm -rf $tmpdir
+	exit 0
+fi
 composer update --no-dev --no-install
 cp composer.lock $composerLock
-rm -rf $tmpdir
 cd $curdir
-nix-update --version=branch=master $ext
+rm -rf $tmpdir
+
+sed -i -e "s/$currentRev/$newRev/g" $fileWithSrc
+nix-update --version=skip $ext
